+++ recipe/joomla.php
@@ -1,0 +1,29 @@
+<?php
+/* (c) Anton Medvedev <anton@medv.io>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Deployer;
+
+require_once __DIR__ . '/common.php';
+
+set('shared_files', ['configuration.php']);
+set('shared_dirs', ['images']);
+set('writable_dirs', ['images']);
+
+task('deploy', [
+    'deploy:info',
+    'deploy:prepare',
+    'deploy:lock',
+    'deploy:release',
+    'deploy:update_code',
+    'deploy:shared',
+    'deploy:writable',
+    'deploy:symlink',
+    'deploy:unlock',
+    'cleanup',
+])->desc('Deploy your project');
+
+after('deploy', 'success');